+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Fix `DuplicateProperty` not reporting duplicate properties in nested rule
   sets
 * Fix `TrailingSemicolon` handling of comma-separated `@import`s
+* Fix `SelectorFormat` to not allow format of pseudo-selectors to be defined,
+  as they are always hyphenated lowercase
 
 ## 0.30.0